# sample-4/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cortex Foundation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "f74109d26aab35daab2d8c617c97c3fa",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/passwordSettings": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Change password",
                "operationId": "0afdbb8f26a5742cd7bf9ace21c5409d",
                "parameters": [
                    {
                        "name": "oldPassword",
                        "in": "query",
                        "description": "Old Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "newPassword",
                        "in": "query",
                        "description": "New Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Password Changed successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register a new user",
                "operationId": "482fd8cbb98dc3225fa0e7ac40c1abe6",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Get logged-in user details",
                "operationId": "5de340fb4e7bf4dc2cd18298959745d3",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/forgot-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Forgot password",
                "operationId": "21e3e385c833b570db9f7fa856264b78",
                "requestBody": {
                    "description": "Blog data including image upload",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "type": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Link sent successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/password/reset": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Reset password",
                "operationId": "12991071ab1240caebde9e735a1ff820",
                "requestBody": {
                    "description": "Blog data including image upload",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "token",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "email",
                                        "type": "email"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "password reset successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}